[FEAT]: 주유내역 CRUD 구현, 코드 정리

+++ CarLog/CarLog/Models/Driving.swift
@@ -11,8 +11,8 @@ struct Driving: Codable {
     var documentID: String?
     let timeStamp: String?
     let id: String?
-    let departDistance: Double?
-    let arriveDistance: Double?
-    let driveDistance: Double?
+    let departDistance: Int?
+    let arriveDistance: Int?
+    let driveDistance: Int?
     let userEmail: String?
 }
